add redux file structure
