+++ client/src/App.jsx
@@ -1,33 +1,12 @@
-import React, { useState } from 'react';
-import MonacoEditor from 'react-monaco-editor';
+import React from 'react';
+import { Editor } from './features';
 
-const App = () => {
-  const [code] = useState('// happy hacking!');
-
-  const onChange = (newValue, e) => {
-    console.log('onChange', newValue, e);
-  };
-
-  const editorDidMount = (editor) => {
-    editor.focus();
-  };
-
-  const options = {
-    selectOnLineNumbers: true,
-  };
-
-  return (
-    <MonacoEditor
-      width="800"
-      height="600"
-      language="javascript"
-      theme="vs-dark"
-      value={code}
-      options={options}
-      onChange={onChange}
-      editorDidMount={editorDidMount}
-    />
-  );
-};
+const App = () => (
+  <>
+    <header>header</header>
+    <Editor />
+    <footer>footer</footer>
+  </>
+);
 
 export default App;
