+++ client/src/features/index.js
@@ -1,0 +1,4 @@
+import Editor from './Editor';
+
+// eslint-disable-next-line import/prefer-default-export
+export { Editor };
